Fix RemoteIndexedMemoryChunk._struct_equals$O (compare place id (int) rather than Place with ==), and performance improvement (use ConcurrentHashMap like GlobalRef).
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@22827 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src-java/x10/runtime/impl/java/PlaceLocalHandle.java
@@ -102,11 +102,11 @@ public final class PlaceLocalHandle<T> implements java.io.Serializable, X10JavaS
     }
 
 	public void $_serialize(X10JavaSerializer serializer) throws IOException {
-		serializer.write(id.longValue());
+		serializer.write((long) id);
 	}
 
 	public static X10JavaSerializable $_deserialize_body(PlaceLocalHandle placeLocalHandle, X10JavaDeserializer deserializer) throws IOException {
-        placeLocalHandle.id = deserializer.readLong();
+        placeLocalHandle.id = (Long) deserializer.readLong();
         return placeLocalHandle;
 	}
